30/09/2023

+++ APP/BUOI3/B2/app/src/main/java/com/example/b2/MainActivity.kt
@@ -41,13 +41,15 @@ class MainActivity : AppCompatActivity() {
         }
 
         btnDelAll.setOnClickListener {
-            while (true){
-                var index = arrFile.size-1
-                val cacheFile = File(cacheDir, arrFile[index])
-                cacheFile.delete()
-                listAdapter.remove(arrFile[index])
-                if(index==0){
-                    break
+            if((arrFile.size)!=0) {
+                while (true) {
+                    var index = arrFile.size - 1
+                    val cacheFile = File(cacheDir, arrFile[index])
+                    cacheFile.delete()
+                    listAdapter.remove(arrFile[index])
+                    if (index == 0) {
+                        break
+                    }
                 }
             }
             listAdapter.notifyDataSetChanged()
